+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 1.1-dev
 * Added warning for setting cookies with empty names
-* Added color differentiation between error messages and warnings  
+* Added color differentiation between error messages and warnings
+* Added detection of non-lowercase function names
 
 ## Version 1.0.2
 * Fixed notice in ```PHP4ConstructorVisitor``` caused by anonymous classes